feat: 받은 지원서 View List 형태로 변경, 받은 지원서 Cell로 분리
